+++ apps/frontend/src/__tests__/api/committee/requestAdd.test.ts
@@ -13,7 +13,7 @@ import {
   createAuthTestSuite,
   type AuthTestConfig,
 } from "../../utils/testUtils";
-import { LEG_DISTRICT_SENTINEL } from "~/lib/constants/committee";
+import { LEG_DISTRICT_SENTINEL } from "@voter-file-tool/shared-validators";
 import {
   mockAuthSession,
   mockHasPermission,
