--- apps/frontend/src/lib/constants/committee.ts
@@ -1,9 +1,0 @@
-/**
- * Committee-related constants
- */
-
-/**
- * Sentinel value used to represent undefined legDistrict in database storage
- * This prevents drift across routes/clients when handling optional legDistrict values
- */
-export const LEG_DISTRICT_SENTINEL = -1;
